LayoutTests/storage/indexeddb/mozilla/key-requirements-put-no-key-expected.txt

Test IndexedDB's behavior puting without key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "key-requirements-put-no-key.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('bar');
Expecting exception from objectStore.put({});
PASS Exception was thrown.
PASS code is 0
PASS ename is 'DataError'
Exception message: DataError: DOM IDBDatabase Exception 0
PASS successfullyParsed is true

TEST COMPLETE
